backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
stockfish = "^3.28.0"
chess = "^1.10.0"
python-dotenv = "^1.0.1"
channels = {extras = ["daphne"], version = "^4.1.0"}
shortuuid = "^1.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
